pyproject.toml
[project]
name = "assignment-vi"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "celery>=5.6.3",
    "chonkie[all]>=1.6.2",
    "chromadb>=1.5.8",
    "fastapi>=0.136.0",
    "mistralai>=2.4.0",
    "pydantic>=2.13.2",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "scalar-fastapi>=1.8.2",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
]

[tool.ruff]
line-length = 99

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
]
